+++ src/main/java/com/blackbuild/groovy/configdsl/transform/ast/DslMethodBuilder.java
@@ -39,6 +39,7 @@ import org.codehaus.groovy.ast.tools.GeneralUtils;
 import java.util.List;
 
 import static com.blackbuild.groovy.configdsl.transform.ast.DSLASTTransformation.NAME_OF_MODEL_FIELD_IN_RW_CLASS;
+import static com.blackbuild.groovy.configdsl.transform.ast.DSLASTTransformation.SET_OWNERS_METHOD;
 import static com.blackbuild.groovy.configdsl.transform.ast.DslAstHelper.hasAnnotation;
 import static org.codehaus.groovy.ast.tools.GeneralUtils.callX;
 import static org.codehaus.groovy.ast.tools.GeneralUtils.ifS;
@@ -90,10 +91,14 @@ public final class DslMethodBuilder extends GenericsMethodBuilder<DslMethodBuild
         return statement(ifS(notX(propX(targetX,"$manualValidation")), callX(targetX, DSLASTTransformation.VALIDATE_METHOD)));
     }
 
-    public DslMethodBuilder optionallySetOwnerOnS(String target, boolean targetHasOwner) {
-        if (!targetHasOwner)
-            return this;
-        return callMethod(varX(target), "set$owner", varX(NAME_OF_MODEL_FIELD_IN_RW_CLASS));
+    public DslMethodBuilder setOwners(String target) {
+        return callMethod(varX(target), SET_OWNERS_METHOD, varX(NAME_OF_MODEL_FIELD_IN_RW_CLASS));
+    }
+
+    public DslMethodBuilder setOwnersIf(String target, boolean apply) {
+        if (apply)
+            return setOwners(target);
+        return this;
     }
 
     public DslMethodBuilder params(Parameter... params) {